==> analysis/04_Biodiversity.Rmd <==
---
title: "Biodiversity analysis"
author: "Calla Bush St George"
date: "`r Sys.Date()`"
output:
  html_document: 
    code_folding: show
    theme: spacelab
    highlight: pygments
    keep_md: no
    toc: yes
    toc_float:
      collapsed: no
      smooth_scroll: yes
      toc_depth: 3
  keep_md: true  
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      fig.path = "../figures/04_Biodiversity/")
```

# Goals

1. Calculate the Hill diversity of the samples.
2. Evaluate the rarefaction curves.
3. Evaluate the diversity values.
4. Make notes of specific samples and their seq depth.

# Setting the Environment

## Set the seed
```{r set-seed}
set.seed(238428)
```


## Load libraries
```{r load-packages}
pacman::p_load(tidyverse, devtools, patchwork, iNEXT, phyloseq,
               install = FALSE)
```

## Load in Data
```{r load-data}
load("data/02_PreProcessing/raw_preprocessed_physeq.RData")
raw_preprocessed_physeq

# Intuition check
min(sample_sums(raw_preprocessed_physeq))

# Make a metadata dataframe
metadata_df <- raw_preprocessed_physeq %>%
  sample_data() %>%
  data.frame()

# Intuition check
head(metadata_df)

# Setting colors for the stations
station_colors <- c(
  "Shipping Channel" = "dodgerblue4",
  "Aransas Bay" = "dodgerblue2",
  "Copano West" = "#D9CC3C",
  "Copano East" = "#A0E0BA",
  "Mesquite Bay" = "#00ADA7")
```


# Diversity Calculations with iNEXT
```{r calc-div}
# Prepare input data
iNEXT_input_df <-
  raw_preprocessed_physeq %>%
  otu_table() %>%
  data.frame()

# Intuition check
dim(iNEXT_input_df)

# Run iNEXT to calculate the hill numbers
# Note: we need to have species in rows and samples in columns
# Remember to set the seed!
#iNEXT_data <-
 # iNEXT(iNEXT_input_df,
 #       q = c(0,1,2), datatype = "abundance")

# Save the file
#save(iNEXT_data, file = "data/04_Biodiversity/iNEXT_data.RData")
```

# Evaluate the diversity
```{r load-div}
load("data/04_Biodiversity/iNEXT_data.RData")
str(iNEXT_data)
```

# Plot Diversity
```{r plot-rarefaction}
# Prepare colors
color_df <- iNEXT_input_df %>%
  colnames() %>%
  data.frame()

# Check
head(color_df)

# Rename the column
colnames(color_df)[1] <- "names"

# Check
head(color_df)

# Make a helper dataframe for plotting with colors
iNext_color_df <-
  color_df %>%
  # Fix the names for merging
   mutate(names = gsub(names, pattern = "[.]", replace = "-"), 
          names = gsub(names, pattern = "X", replace = "")) %>%
  # Merge with metadata
  left_join(metadata_df, by = "names") %>%
  # Merge with colors for plotting with ggiNEXT
  left_join(data.frame(station_colors = station_colors,
            station = names(station_colors)),
            by = "station") 
```

# Plot Rarefaction with `ggiNEXT`

```{r ggiNEXT, fig.width=8, fig.height=3.5}
# Plot rarefaction!
# Rarefaction/extrapolation curve, type = 1

# Order q:
 # 0 = Richness or total number of taxa
 # 1 = Exponential shannon or number of "common" taxa
 # 2 = Inverse simpson, number of dominant taxa

ggiNEXT(iNEXT_data, type = 1, facet.var = "Order.q") + 
  facet_wrap(~Order.q, scales = "fixed") +
  scale_color_manual(values = iNext_color_df$station_colors, guide = FALSE) +
  scale_fill_manual(values = iNext_color_df$station_colors, guide = FALSE) +
  scale_shape_manual(values = base::rep(17, nsamples(raw_preprocessed_physeq)),
                     guide = FALSE) +
  theme(legend.position = "none")
```

# Manually plot Diversity

## Rarefaction
```{r iNEXT-manual}
iNEXT_manual_df <- 
  iNEXT_data$iNextEst$size_based %>%
  # $ is saying within the iNEXT_data list we want the iNEXTEst, then within
  # that we want the size_based list
  dplyr::rename(names = Assemblage) %>%
  # Fix sample names
  mutate(names = gsub(names, pattern = "[.]", replace = "-"),
         names = gsub(names, pattern = "X", replace = "")) %>%
  # Join with metadata
  left_join(., metadata_df, by = "names") %>%
  # Add colors to data frame
  left_join(., data.frame(station_colors = station_colors,
                          station = names(station_colors)),
            by = "station")

# Inspect 
dim(iNEXT_manual_df)
str(iNEXT_manual_df)

# Plot it - Rarefaction Curve
iNEXT_manual_df %>%
  # Filter out rows that are calculated by rarefaction from iNEXT
  dplyr::filter(Method == "Rarefaction") %>%
  # Make the actual rarefaction plot with
  # the number of sequences on the x-axis and diversity on the y-axis
  ggplot(aes(x = m, y = qD, color = station, group = names)) + 
  geom_line() +
  #geom_point()
  # Challenge: facet with the stations
  scale_color_manual(values = station_colors) +
  facet_grid(Order.q~station, scales = "fixed") +
  theme(legend.position = "bottom")

# Plot it - Extrapolation 
iNEXT_manual_df %>%
  # Filter out rows that are calculated by rarefaction from iNEXT
  dplyr::filter(Method == "Extrapolation") %>%
  # Make the actual rarefaction plot with
  # the number of sequences on the x-axis and diversity on the y-axis
  ggplot(aes(x = m, y = qD, color = station, group = names)) + 
  geom_line() +
  #geom_point()
  scale_color_manual(values = station_colors) +
  facet_grid(Order.q~station, scales = "fixed") +
  theme(legend.position = "bottom")
```

# Diversity versus salinity

## Environmental variable check
```{r environmental-pairs, fig.width=8, fig.height=8}
head(metadata_df)
# Pull out environmental variables
env_df <-
  metadata_df %>%
  dplyr::select(names, water_tempC:DO_mgL)
# inspect
head(env_df)

# plot the correlations
pairs(dplyr::select(env_df, -names), upper.panel = NULL)
```

## Diversity versus salinity
```{r div-vs-salinity, fig.height=3.5, fig.width=6}
# PSU = practial salinity unit
# 1 PSU = 1 gram of salt per 1,000 grams of water
# 1 PSU = 1 ppt (part per thousand)
# Freshwater = 0 - 0.5
# Brackish = 0.5 - 30
# Marine = >30

iNEXT_manual_df %>%
  dplyr::filter(Method == "Observed") %>%
  ggplot(aes(x = salinity_psu, y = qD)) +
  geom_point(aes(color = station)) +
  stat_smooth(method = "lm", formula = y ~poly(x,2)) + 
  facet_wrap(.~Order.q, scales = "free") +
  theme(legend.position = "bottom") +
  scale_color_manual(values = station_colors) + 
  labs(x = "Salinity (PSU)", y = "# of ASVs")

```




##Session information
```{r session-info}
#Ensure reproducibility
devtools::session_info()
```
